Added variables, changed f64 to f32

+++ src/foreign/Numbers.js
@@ -1,5 +1,5 @@
 "use strict";
 
 exports.toBytes = function(n) {
-    return Array.from(new Uint8Array(new Float64Array([n]).buffer));
+    return Array.from(new Uint8Array(new Float32Array([n]).buffer));
 };
